multiples nombres y pasies
