+++ ejercicio 1/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+    <link rel="stylesheet" href="style.css">
+  </head>
+  <div class="contenedor"></div>
+   <span class="verticalLine"></span>
+   <span class="horizontalLine"></span>
+
+</html>
