+++ ejercicio 1/style.css
@@ -1,0 +1,28 @@
+.contenedor {
+  /* width: 100%;
+   height: 480px;
+   position: relative;
+   background-color: red; */
+}
+.verticalLine {
+   display: block;
+   position: absolute;
+   background-color: gray;
+   width: 2px;
+   top: 0px;
+   bottom: 0px;
+   left: 50%;
+   height: 100vh;
+}
+
+.horizontalLine {
+   position: absolute; 
+   background-color: gray; 
+   height: 100vw; 
+   width: 2px;
+   top: 50%;
+   bottom: 0;
+   left: 50%;
+   transform: translate(-50%, -50%) rotate(90deg);
+  
+}
